refactor: apply pyflakes checks
This commit applies pyflakes checks to the codebase and removes a large number of unused variables and imports.
Notable non-automatic fixes in this commit include removing variables that are never checked while testing error conditions.
This also removes an unused invocation of consensus during the test_aggregate set of tests that appears to be left over from copy and pasting tests.