refactor(docs): apply flake8-builtins checks
This commit applies the flake8 builtins checks that checks for variables shadowing python built-in functions and variables. As part of this, we rename the docs variable `copyright` to the alias `project_copyright` to avoid shadowing the python built-in function. https://www.sphinx-doc.org/en/master/usage/configuration.html#confval-project_copyright